refactor(linux): Extract widgets to separate classes and files
